KonuTahmin/NLP/requirements.txt
antlr4-python3-runtime==4.8
beautifulsoup4==4.11.1
certifi==2021.10.8    
charset-normalizer==2.0.12
google==3.0.0
idna==3.3
joblib==1.1.0
numpy==1.19.0
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
urllib3==1.26.9
zemberek-python==0.1.2